Added check that all diffs are listed in the generate-upstream-patches.filelist.

+++ plo/generate-upstream-patches.sh
@@ -18,6 +18,22 @@ if [[ -z "${BASEINFO}" ]]; then
     exit 1
 fi
 
+mapfile -t git_reported_diff_files < <( git diff --numstat ${BASEREV} -- llvm/ compiler-rt/ clang/ lld/ | tr "\t" " " | tr -s " " | cut -f3 -d " " )
+for GF in "${git_reported_diff_files[@]}" ; do
+    # Try to find GF in ALL_FILES[@]
+    FOUND=
+    for AF in "${ALL_FILES[@]}" ; do
+        if [[ "$GF" == "$AF" ]]; then
+            FOUND=1
+            break
+        fi
+    done
+    if [[ -z "${FOUND}" ]]; then
+        echo "Git reported \"$GF\" is not in \"generate-upstream-patches.filelist\", please update the file."
+        exit 1
+    fi
+done
+
 echo "Generate patches based on: ${BASEINFO}"
 
 if [[ "$(( ${#ALL_FILES[@]} % 2 ))" != "0" ]]; then
@@ -46,13 +62,17 @@ declare -A component_map
 while ((idx<"${#ALL_FILES[@]}")); do
     component="${ALL_FILES[$idx]}"
     file="${ALL_FILES[$((idx+1))]}"
-    patch_file="${DDIR}/upstream-${component}.patch"
-    if [[ -z "${component_map[$component]}" ]]; then
-        rm -f "${patch_file}"
-        component_map[$component]="${patch_file}"
-        echo "Creating patch: ${patch_file}"
+    if [[ "${component}" == "/dev/null" ]]; then
+        patch_file=${component}
+    else
+        patch_file="${DDIR}/upstream-${component}.patch"
+        if [[ -z "${component_map[$component]}" ]]; then
+            rm -f "${patch_file}"
+            component_map[$component]="${patch_file}"
+            echo "Creating patch: ${patch_file}"
+        fi
+        git diff --unified=99999 "${BASEREV}" -- ${file} >> ${patch_file}
     fi
-    git diff --unified=99999 "${BASEREV}" -- ${file} >> ${patch_file}
     idx="$((idx+2))"
 done
